% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_Z_mj.R
\name{infer_Z_mj}
\alias{infer_Z_mj}
\title{infer_Z_mj}
\usage{
infer_Z_mj(X, seq_depth = "TS")
}
\arguments{
\item{X}{matrix of n by p data where n are observations and p are numeric variables}

\item{seq_depth}{specify "TS" for raw count data for centralized log-normalizing process
or or "unif" for empirical cumulative distribution function method. See Prost paper for details.}
}
\value{
Empirical correlation matrix with same dimensions as input.
}
\description{
Zero inflated empirical correlation matrix
(ver. 20190920)
}
\examples{

\dontrun{
res <- infer_Z_mj(data(as.matrix(counts_el)))
 }

}
\author{
Margaret Janiczek
}
